+++ Lib/test/libregrtest/runtest.py
@@ -143,6 +143,7 @@ SPLITTESTDIRS = {
     "test_asyncio",
     "test_concurrent_futures",
     "test_future_stmt",
+    "test_gdb",
     "test_multiprocessing_fork",
     "test_multiprocessing_forkserver",
     "test_multiprocessing_spawn",
