+++ Makefile.pre.in
@@ -1966,6 +1966,7 @@ TESTSUBDIRS=	ctypes/test \
 		test/test_email \
 		test/test_email/data \
 		test/test_future_stmt \
+		test/test_gdb \
 		test/test_import \
 		test/test_import/data \
 		test/test_import/data/circular_imports \
