+++ Misc/NEWS.d/next/Tests/2023-09-28-12-25-19.gh-issue-109972.GYnwIP.rst
@@ -1,0 +1,2 @@
+Split test_gdb.py file into a test_gdb package made of multiple tests, so tests
+can now be run in parallel. Patch by Victor Stinner.
